fix: when updating a contact information, the updated data is not reflected on the contact
